+++ Prog/BookGen/CommandArguments/IconSearchArguments.cs
@@ -7,14 +7,15 @@ namespace BookGen.CommandArguments;
 
 internal class IconSearchArguments : SearchArgumentsBase
 {
-    
+    [Switch("i8", "icons8")]
     public bool? Icons8 { get; set; }
+
+    [Switch("svg", "svgrepo")]
     public bool? SvgRepo { get; set; }
-    public bool? MaterialDesign { get; set; }
 
     public bool All
     {
-        get => (!Icons8.HasValue && !SvgRepo.HasValue && !MaterialDesign.HasValue)
-            || (Icons8 == true && SvgRepo == true && MaterialDesign == true);
+        get => (!Icons8.HasValue && !SvgRepo.HasValue)
+            || (Icons8 == true && SvgRepo == true);
     }
 }